+++ CHANGELOG.md
@@ -3,5 +3,5 @@ All notable changes to this project will be documented in the changelog of the r
 This project follows the [Semantic Versioning standard](https://semver.org/).
 
 For the changelogs of the crates see:
-- okapi-operation: [./okapi-operation/CHANGELOG.md](./okapi-operation/CHANGELOG.md)
-- okapi-operation-macro: [./okapi-operation-macro/CHANGELOG.md](./okapi-operation-macro/CHANGELOG.md)
+- speka: [./speka/CHANGELOG.md](./speka/CHANGELOG.md)
+- speka-macro: [./speka-macro/CHANGELOG.md](./speka-macro/CHANGELOG.md)